# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HelpDesk API (AS-IS)
  version: 0.1.0
  description: OpenAPI baseline reflecting currently implemented routes.
servers:
  - url: https://example.com
    description: Replace with deployment host
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    ErrorAsIs:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [REQUESTER, AGENT, ADMIN]
        organizationId:
          type: string
          format: uuid
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ticketCount:
          type: integer
          description: Total tickets created by this user
        activeTicketCount:
          type: integer
          description: Active tickets assigned to this user
      additionalProperties: true
...
    AdminAuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          enum: [CREATE, UPDATE, DELETE]
        resource:
          type: string
          enum: [USER, TEAM, TAG, SLA]
        resourceId:
          type: string
        actorId:
          type: string
          format: uuid
        actor:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
              nullable: true
            email:
              type: string
              format: email
        data:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    PageInfo:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
...
    TeamMembership:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserSummary'
        joinedAt:
          type: string
          format: date-time
...
paths:
...
  /api/admin/audit-events:
    get:
      summary: List admin audit events
      description: Returns paginated admin audit events for the organization. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: resourceId
          in: query
          schema:
            type: string
          description: Filter by resource ID
        - name: actorId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by actor user ID
      responses:
        '200':
          description: Audit events list
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminAuditEvent'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
...
  /api/admin/teams/{id}/memberships:
    get:
      summary: List team memberships
      description: Returns all memberships for a specific team. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team memberships
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberships:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamMembership'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    post:
      summary: Add team member
      description: Adds a user to a team. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  membership:
                    $ref: '#/components/schemas/TeamMembership'
        '400':
          description: Validation error or user already in team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    delete:
      summary: Remove team member
      description: Removes a user from a team. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team or membership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
...